Feature/user graphql

See merge request s02-final/s02p31a404!30

+++ server/market/src/main/java/com/ssafy/market/domain/file/resolver/FileMutation.java
@@ -8,6 +8,7 @@ import com.ssafy.market.domain.file.dto.UpdateFileInput;
 import com.ssafy.market.domain.file.repository.FileRepository;
 import com.ssafy.market.domain.product.domain.Product;
 import com.ssafy.market.domain.product.repository.ProductRepository;
+import com.ssafy.market.global.exception.DomainNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import javax.transaction.Transactional;
@@ -21,20 +22,30 @@ public class FileMutation implements GraphQLMutationResolver {
 
     @Transactional
     public FileOutput createFile(CreateFileInput input){
-        Product product = productRepository.findById(input.getProductId()).get();
+        Product product = productRepository.findByProductId(input.getProductId());
+        if(product==null){
+            throw new DomainNotFoundException("productId",input.getProductId());
+        }
         File file = fileRepository.save(new File(null, product,input.getImgPath()));
         FileOutput output = new FileOutput(file.getFileId(), product.getProductId(), input.getImgPath());
         return output;
     }
     @Transactional
     public FileOutput updateFile(UpdateFileInput input){
-        File file = fileRepository.findById(input.getFileId()).get();
+        File file = fileRepository.findByFileId(input.getFileId());
+        if(file==null){
+            throw new DomainNotFoundException("fileId",input.getFileId());
+        }
         file.update(input.getImgPath());
         FileOutput output = new FileOutput(input.getFileId(),file.getProduct().getProductId(),input.getImgPath());
         return output;
     }
     @Transactional
     public int deleteFile(Long id){
+        File file = fileRepository.findByFileId(id);
+        if(file==null){
+            throw new DomainNotFoundException("fileId " , id);
+        }
         return fileRepository.deleteByFileId(id);
     }
 }
